+++ packages/rspeedy/core/src/config/rsbuild/index.ts
@@ -17,7 +17,9 @@ export function toRsbuildConfig(
 ): UndefinedOnPartialDeep<RsbuildConfig> {
   return {
     dev: {
+      hmr: config.dev?.hmr ?? true,
       lazyCompilation: false,
+      liveReload: config.dev?.liveReload ?? true,
       watchFiles: config.dev?.watchFiles,
       // We expect to use different default writeToDisk with Rsbuild
       writeToDisk: config.dev?.writeToDisk ?? true,