+++ packages/rspeedy/core/test/plugins/dev.plugin.test.ts
@@ -154,16 +154,13 @@ describe('Plugins - Dev', () => {
     await rsbuild.createDevServer({ compiler })
 
     // See: https://github.com/web-infra-dev/rsbuild/pull/2303
-    expect(compiler.hooks.compile.taps.map(i => i.name)).toMatchInlineSnapshot(`
-      [
-        "rsbuild-dev-server",
-      ]
-    `)
+    expect(compiler.hooks.compile.taps.map(i => i.name)).toMatchInlineSnapshot(
+      `[]`,
+    )
 
     expect(compiler.hooks.invalid.taps.map(i => i.name)).toMatchInlineSnapshot(`
       [
         "rsbuild-dev-server",
-        "rsbuild-dev-middleware",
       ]
     `)
 
@@ -455,10 +452,48 @@ describe('Plugins - Dev', () => {
     )
   })
 
+  test('environment.dev.hmr: false', async () => {
+    const rsbuild = await createStubRspeedy({
+      environments: {
+        lynx: {
+          dev: {
+            hmr: false,
+          },
+        },
+      },
+    })
+
+    const config = await rsbuild.unwrapConfig()
+
+    expect(config.resolve?.alias).toHaveProperty(
+      '@lynx-js/webpack-dev-transport/client',
+      expect.stringContaining('hot=false'),
+    )
+  })
+
   test('dev.hmr: true', async () => {
     const rsbuild = await createStubRspeedy({
       dev: {
         hmr: true,
+      },
+    })
+
+    const config = await rsbuild.unwrapConfig()
+
+    expect(config.resolve?.alias).toHaveProperty(
+      '@lynx-js/webpack-dev-transport/client',
+      expect.stringContaining('hot=true'),
+    )
+  })
+
+  test('environment dev.hmr: true', async () => {
+    const rsbuild = await createStubRspeedy({
+      environments: {
+        lynx: {
+          dev: {
+            hmr: true,
+          },
+        },
       },
     })
 
@@ -496,10 +531,48 @@ describe('Plugins - Dev', () => {
     )
   })
 
+  test('environment dev.liveReload: false', async () => {
+    const rsbuild = await createStubRspeedy({
+      environments: {
+        lynx: {
+          dev: {
+            liveReload: false,
+          },
+        },
+      },
+    })
+
+    const config = await rsbuild.unwrapConfig()
+
+    expect(config.resolve?.alias).toHaveProperty(
+      '@lynx-js/webpack-dev-transport/client',
+      expect.stringContaining('live-reload=false'),
+    )
+  })
+
   test('dev.liveReload: true', async () => {
     const rsbuild = await createStubRspeedy({
       dev: {
         liveReload: true,
+      },
+    })
+
+    const config = await rsbuild.unwrapConfig()
+
+    expect(config.resolve?.alias).toHaveProperty(
+      '@lynx-js/webpack-dev-transport/client',
+      expect.stringContaining('live-reload=true'),
+    )
+  })
+
+  test('environments dev.liveReload: true', async () => {
+    const rsbuild = await createStubRspeedy({
+      environments: {
+        lynx: {
+          dev: {
+            liveReload: true,
+          },
+        },
       },
     })