+++ packages/rspeedy/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @lynx-js/rspeedy
 
+## 0.11.6
+
+### Patch Changes
+
+- Should apply `dev.hmr` and `dev.liveReload` to Rsbuild config. ([#1882](https://github.com/lynx-family/lynx-stack/pull/1882))
+
+- Support CLI flag `--root` to specify the root of the project. ([#1836](https://github.com/lynx-family/lynx-stack/pull/1836))
+
+## 0.11.5
+
+### Patch Changes
+
+- Bump Rsbuild v1.5.13 with Rspack v1.5.8. ([#1849](https://github.com/lynx-family/lynx-stack/pull/1849))
+
 ## 0.11.4
 
 ### Patch Changes