+++ packages/rspeedy/plugin-react-alias/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @lynx-js/react-alias-rsbuild-plugin
 
+## 0.11.2
+
+## 0.11.1
+
 ## 0.11.0
 
 ## 0.10.14