+++ packages/web-platform/web-rsbuild-server-middleware/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @lynx-js/web-rsbuild-server-middleware
 
+## 0.18.0
+
+## 0.17.2
+
 ## 0.17.1
 
 ### Patch Changes